Add activity names to individual activity mappings
The titles of weekly events (Nightfall and Heroics) are not the actual
strike names, which can cause confusion for callers.

This change adds the strike's common name to the dictionary returned
from processing the nightfall or heroic. The daily strike is correctly
set, so the name will match the title in order to be consistent.

The crucible also uses different activity titles for its daily
activities, but we do not currently have a mapping for this, so the name
attribute is set to the title.